test_api.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()

api_key = os.getenv('GEMINI_API_KEY')
print(f"API Key found: {bool(api_key)}")
if not api_key: 
    print("No API key found!")
    exit(1)

genai.configure(api_key=api_key)

try:
    model = genai.GenerativeModel('gemini-1.5-flash')
    response = model.generate_content("Hello, this is a test.")
    print(f"SUCCESS: Generated content: {response.text}")
except Exception as e:
    print(f"FAILURE: {e}")
