requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
PyMuPDF==1.23.8
groq==0.5.0
reportlab==4.0.7
xhtml2pdf==0.2.14

